изменен тип аргумента метода get()

+++ src/main/java/database/dao/UserDAO.java
@@ -47,7 +47,7 @@ class UserDAO {
         session.getTransaction().commit();
     }
 
-    User get(int id) {
+    User get(long id) {
         Session session = HibernateUtil.getSessionFactory().getCurrentSession();
         session.getTransaction().begin();
